feat: protect all peers in low buckets, tag everyone else with 5
Anything else is easy to game. This way, we remain connected to the core peers
and slightly prefer all others.

Pruning doesn't really matter as we'll just reconnect. The only concern is that
this will lock in 40 connections. However, that's nothing for libp2p anyways.
and the DHT will create that many connections _just_ for a single query.