increase provider maddr ttl
We keep the Multiaddresses of providing peers around for much longer
as this means we'll return them alongside the provider records and
likely make the second DHT lookup for the peer record obsolete.

The assumption is that peers don't change network addresses often.
For context: https://github.com/ipfs/kubo/issues/9264