feat: add option to skip sending GetProviders closer peers
This is to reduce the amount of outbound traffic from Hydras boosters,
which significantly reduces operating costs. This really should only
be used on the public network in very specific situations, such as
when the Hydra node is not returning useful DHT information to begin
with, but is only being used to bridge the network with
indexers, which is only temporarily a valid use case for the DHT until
delegated routing is finalized and released. Once that happens, we
should remove this ability and instead direct users to use delegated
routing instead of relying on Hydra nodes to bridge content routers
with the DHT.