Allow DHT crawler to be swappable
The accelerated DHT client uses a fixed implementation of DHT crawler
with a parallelism hardcoded to 200. This offers a reasonable general
case default behaviour but limits experimentation with alternative
crawling strategies with varied or dynamic degree of parallelism.

To avoid this the change here:
 * refactor the existing `Crawler` to `DefaultCrawler`
 * introduce `Crawler` as interface type, and
 * add option to set the `Crawler` in the accelerated DHT client with
  fallback to `DefaultCrawler` at parallelism of 200.

As a result the net functionality of the code remains unchanged while it
allows the crawling logic to be swapped entirely with an alternative
implementation.

Additionally, the changes here parameterize the dial addr extend
duration with fallback to the previously hardcoded value of 30 minutes.