fix(crawler): separate timeout per query
This aims to correctly apply intention of fix from
https://github.com/libp2p/go-libp2p-kad-dht/pull/996

It did not work correctly because `context.WithTimeout`  was created once
per goroutine but was used for multiple queries within
the `for p := range jobs` loop.

This meant once the timeout expires for the first query, all subsequent
queries in that worker would immediately fail since they're using the
same expired context.