fix(provider): resume cycle from persisted keystore (#1193)
* refactor(keystore): keep track of size

* pkg godoc

* Keystore default path

Update to "keystore" only, since it should be a namespace of the
"provider" datastore defined in provider.go.

* persist size on keystore.Close()

* deduplicate put to altDs

* tests

* adding logger to Keystore

* don't refresh size after Reset() failure

If there was an error when writing to the alt datastore during the
Reset() operation, Reset() is cancelled, and alt datastore is emptied.
When Reset() is called again, we start with a fresh alt datastore. Hence
it isn't useful to update the alt keystore size after put failure.

* log refreshSize failures

* fix concurrency in test

If Put() happen concurrently to reset, they *may* be processed after the
reset is over, and it doesn't have an impact on the size. Hence, we can
sleep for a bit to let the keystore process the Put() operations
potentially happening after Reset, before getting the size.

* initialize logger

* fix: flaky TestKeystoreResetSizeWithConcurrentPuts

* address review

* test resume cycle

* ResettableKeystore: persist active datastore index

* catchupPendingWork after regions added to reprovide queue

* ResettableKeystore tests

* fix: datastore crash when emptying datastore after close

* tests

* prevent writing to datastore after close

* cancel network operations on close